official_paddleocr_serving/requirements_official.txt
paddleocr[all]
paddlex
requests
pillow
opencv-python
numpy
